Trying to fix a display issue on Linux, part 2
This is a disgusting hackery mess of an attempt at setting up a hidden window on Linux. On XFCE (possibly others), I'm unable to set the initial position outside the visible window. I know, I know, why would you ever want to do that? :P

I need to see how this looks on Mac, as well.